requirements.txt
dagster==1.10.0
pandas==2.2.2
dagster-webserver==1.10.0
requests==2.31.0
python-dotenv==1.0.1
pre-commit==4.1.0